Complete UserAsset separate component
